feat(nenyr-parser): add Animation parser with error and suggestion messages
- Implemented the **Animation parser** into the **NenyrParser**, fully functional for parsing animations defined in the Nenyr document.
- Added comprehensive **error** and **suggestion messages** to the Animation parser to improve feedback and debugging.

Next steps:
  - Complete **documentation** and **testing** for the Animation parser to finalize it.
  - Begin implementation of the **Variables parser** after completing the Animation parser.

+++ src/tokens/mod.rs
@@ -39,7 +39,7 @@ pub enum NenyrTokens {
     // Value collectors
     //Unknown(char),
     StringLiteral(String),
-    Number(i64),
+    Number(f64),
     Identifier(String),
 
     // Syntax tokens
